--- lab_2/rows.json
@@ -1,4 +1,0 @@
-{
-    "c++": " ",
-    "java": " "
-}
