+++ lab_2/task1/cpp_row.txt
@@ -1,0 +1,1 @@
+111001011000110001001111100111100101000011000000100000110101101110010111011010100110100000110101111100100001010111101101000011000
