+++ lab_2/task1/java_row.txt
@@ -1,0 +1,1 @@
+10101100000001010000111001011111111000011010000110100110101111101100111000001001011110101001110001001001010110110111100000100010
